"Final Api"

+++ app/Http/Controllers/employee/EmployeeController.php
@@ -3,45 +3,61 @@
 namespace App\Http\Controllers\employee;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\api\StoreEmployeeRequest;
+use App\Http\Traits\GeneralTrait;
+use App\Http\Traits\imageTrait;
 use App\Models\Employe;
 use App\Models\Section;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
 
 class EmployeeController extends Controller
 {
+    use GeneralTrait;
+    use imageTrait;
 
     public function index()
     {
-        $employees=Employe::all();
+        $employees=Employe::where('club_id',Auth::user()->club_id)->get();
         return view('pages.employee.index' , compact('employees'));
     }
 
 
     public function create()
     {
-        $sections=Section::all();
+        $sections=Section::where('club_id',Auth::user()->club_id)->get();
         return view('pages.employee.create',compact('sections'));
     }
 
     public function store(StoreEmployeeRequest $request)
     {
+        DB::beginTransaction();
         try {
-
             $employees  = new Employe();
+            $employees->club_id = Auth::user()->club_id;
             $employees->name = ['en' => $request->name_en, 'ar' => $request->name_ar];
             $employees->email = $request->email;
             $employees->password = Hash::make($request->password);
-            $employees->number = $request->number;
             $employees->section_id = $request->section_id;
             $employees->description = $request->description;
             $employees->full_description = $request->full_description;
             $employees->date_of_birth = $request->date_of_birth;
             $employees->emp_id = $request->emp_id;
+            $employees->emp_status = $request->emp_status;
+            $employees->national_id = $request->national_id;
             $employees->save();
+            if ($request->hasfile('image_path')) {
+                $employee_image = $this->saveImage($request->image_path, 'attachments/employees/' .Auth::user()->club_id.'/'. $employees->id);
+                $employees->image_path = $employee_image;
+                $employees->save();
+            }
+            DB::commit();  // insert data
+
             session()->flash('Add', trans('notifi.add'));
             return redirect()->route('employee.index');
         } catch (\Exception $e) {
+            DB::rollback();
             return redirect()->back()->withErrors(['error' => $e->getMessage()]);
         }
     }
@@ -55,27 +71,32 @@ class EmployeeController extends Controller
 
     public function edit($id)
     {
-        $employees = Employe::findorfail($id) ;
-        $sections=Section::all();
+        $employees = Employe::where('club_id',Auth::user()->club_id)->findorfail($id) ;
+        $sections=Section::where('club_id',Auth::user()->club_id)->get();
         return view('pages.employee.edit', compact('employees','sections'));
     }
 
     public function update(Request $request, $id)
     {
         try {
-
-
-            $employees =Employe::findorfail($request->id);
+            $employees =Employe::where('club_id',Auth::user()->club_id)->find($request->id);
             $employees->name = ['en' => $request->name_en, 'ar' => $request->name_ar];
             $employees->email = $request->email;
             $employees->password = Hash::make($request->password);
-            $employees->number = $request->number;
             $employees->section_id = $request->section_id;
             $employees->description = $request->description;
             $employees->full_description = $request->full_description;
             $employees->date_of_birth = $request->date_of_birth;
             $employees->emp_id = $request->emp_id;
+            $employees->emp_status = $request->emp_status;
+            $employees->national_id = $request->national_id;
             $employees->save();
+            if ($request->hasfile('image_path')) {
+                $this->deleteFile('employees',$request->id);
+                $_image = $this->saveImage($request->image_path, 'attachments/employees/' .Auth::user()->club_id.'/'. $food->id);
+                $food->image_path = $_image;
+                $food->save();
+            }
             session()->flash('update', trans('notifi.update'));
             return redirect()->route('employee.index');
         } catch (\Exception $e) {
@@ -87,7 +108,7 @@ class EmployeeController extends Controller
     public function destroy(Request $request)
     {
         try {
-            Employe::destroy($request->id);
+            Employe::where('club_id',Auth::user()->club_id)->destroy($request->id);
             session()->flash('delete', trans('notifi.delete'));
             return redirect()->route('employee.index');
         } catch (\Exception $e) {
